Move backward compat tests into their own java package
Master Issue: #1030
